gpu: Create gpu_record.cpp and gpu_setup.cpp
Cleanup gpu_validation.cpp by creating gpu_record.cpp for Record
functions, and gpu_setup.cpp for validation related state objects
creation/destruction
Reorganise function declarations in gpuav::Validator accordingly
