+++ BUILD.gn
@@ -167,6 +167,8 @@ vvl_sources = [
   "layers/gpu_validation/gpu_state_tracker.h",
   "layers/gpu_validation/gpu_subclasses.cpp",
   "layers/gpu_validation/gpu_subclasses.h",
+  "layers/gpu_validation/gpu_record.cpp",
+  "layers/gpu_validation/gpu_setup.cpp",
   "layers/gpu_validation/gpu_validation.cpp",
   "layers/gpu_validation/gpu_validation.h",
   "layers/gpu_validation/gpu_vuids.cpp",
