+++ layers/CMakeLists.txt
@@ -198,11 +198,13 @@ target_sources(vvl PRIVATE
     gpu_validation/gpu_state_tracker.h
     gpu_validation/gpu_subclasses.cpp
     gpu_validation/gpu_subclasses.h
-    gpu_validation/gpu_validation.cpp
+    gpu_validation/gpu_record.cpp
+    gpu_validation/gpu_setup.cpp
     gpu_validation/gpu_validation.cpp
     gpu_validation/gpu_validation.h
     gpu_validation/gpu_vuids.cpp
     gpu_validation/gpu_vuids.h
+    gpu_shaders/gpu_shaders_constants.h
     object_tracker/object_lifetime_validation.h
     object_tracker/object_tracker_utils.cpp
     state_tracker/base_node.cpp
